Dns peer feed no queue (#3716)
* node sources

* tracking Goerli sessions

* DNS peering working

* added PeerPool mock for tests

* changed the Node IDs in tests of discovery so they no longer conflict with master node

* fixed tests by starting peer pool and returning static nodes properly from the mock - also found one bug with session that was fixed by the changes

* test fixes

* Review changes

* comment

* use string.Empty where possible

* Move enr discovery to lower allocation code.

* Make EnrDiscoveryTests runnable - it fails now

* refactor, tests and fixing padding

* Further changes

* Make tests explicit as CI blocks DNS

+++ src/Nethermind/Nethermind.Network.Discovery/Messages/PingMsg.cs
@@ -33,7 +33,7 @@ public class PingMsg : DiscoveryMsg
     /// <summary>
     /// https://eips.ethereum.org/EIPS/eip-868
     /// </summary>
-    public int? EnrSequence { get; set; }
+    public long? EnrSequence { get; set; }
 
     public PingMsg(PublicKey farPublicKey, long expirationTime, IPEndPoint source, IPEndPoint destination, byte[] mdc)
         : base(farPublicKey, expirationTime)
